# docs/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/activity": {
      "get": {
        "operationId": "listActivity",
        "parameters": [
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "eventType",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "profile_synced",
                "game_completed",
                "rare_achievement_synced",
                "guide_published",
                "guide_commented",
                "guide_voted",
                "session_created",
                "session_joined",
                "session_commented",
                "milestone_reached"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityFeedResponseDto"
                }
              }
            }
          }
        },
        "summary": "List latest public activity events",
        "tags": [
          "activity"
        ]
      }
    },
    "/profiles/{steamId}/activity": {
      "get": {
        "operationId": "listProfileActivity",
        "parameters": [
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "steamId",
            "required": true,
            "in": "path",
            "description": "Steam 64-bit profile id",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityFeedResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Profile was not found."
          }
        },
        "summary": "List public activity for a Steam profile",
        "tags": [
          "activity"
        ]
      }
    },
    "/games/{steamAppId}/activity": {
      "get": {
        "operationId": "listGameActivity",
        "parameters": [
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "steamAppId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityFeedResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Game was not found."
          }
        },
        "summary": "List public activity for a Steam game",
        "tags": [
          "activity"
        ]
      }
    },
...
  },
  "info": {
    "title": "Steam Achievement Tracker API",
    "description": "API for syncing and reading Steam achievement intelligence data",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Docker backend"
    }
  ],
  "components": {
    "schemas": {
...
      "ActivityActorResponseDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "nullable": true,
            "example": "Steam Hunter"
          },
          "steamId": {
            "type": "string",
            "nullable": true,
            "example": "76561198000000000"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true,
            "example": null
          },
          "publicSlug": {
            "type": "string",
            "nullable": true,
            "example": "nero"
          }
        }
      },
      "ActivitySteamProfileResponseDto": {
        "type": "object",
        "properties": {
          "steamId": {
            "type": "string",
            "example": "76561198000000000"
          },
          "personaName": {
            "type": "string",
            "nullable": true,
            "example": "Steam Hunter"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true,
            "example": null
          }
        },
        "required": [
          "steamId"
        ]
      },
      "ActivityEventResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "3dd0928a-28f6-4e9b-a5ad-164d536b8d95"
          },
          "eventType": {
            "type": "string",
            "enum": [
              "profile_synced",
              "game_completed",
              "rare_achievement_synced",
              "guide_published",
              "guide_commented",
              "guide_voted",
              "session_created",
              "session_joined",
              "session_commented",
              "milestone_reached"
            ],
            "example": "guide_published"
          },
          "occurredAt": {
            "type": "string",
            "example": "2026-05-17T12:00:00.000Z"
          },
          "actor": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/ActivityActorResponseDto"
              }
            ]
          },
          "steamProfile": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/ActivitySteamProfileResponseDto"
              }
            ]
          },
          "steamAppId": {
            "type": "number",
            "nullable": true,
            "example": 910001
          },
          "entityType": {
            "type": "string",
            "example": "guide"
          },
          "entityId": {
            "type": "string",
            "nullable": true,
            "example": "3dd0928a-28f6-4e9b-a5ad-164d536b8d95"
          },
          "metadata": {
            "type": "object",
            "example": {
              "title": "Demo Roadmap"
            }
          }
        },
        "required": [
          "id",
          "eventType",
          "occurredAt",
          "entityType",
          "metadata"
        ]
      },
      "ActivityFeedResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActivityEventResponseDto"
            }
          },
          "total": {
            "type": "number",
            "example": 30
          },
          "limit": {
            "type": "number",
            "example": 30
          },
          "offset": {
            "type": "number",
            "example": 0
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ]
      },
...
    }
  }
}